Move AutoFilter reading to separate class
The AutoFilter is used in 3 places:
* Sheet
* Table
* Pivot table

Therefore it is better to have it in a separate place.
